fix warnings (mostly %d vs. %ld)
git-svn-id: https://svn.osgeo.org/gdal/trunk@14430 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/ogr/ogrsf_frmts/tiger/tigerfilebase.cpp
@@ -223,7 +223,7 @@ void TigerFileBase::EstablishFeatureCount()
         CPLError( CE_Warning, CPLE_FileIO,
                   "TigerFileBase::EstablishFeatureCount(): "
                   "File length %d doesn't divide by record length %d.\n",
-                  nFileSize, nRecordLength );
+                  (int) nFileSize, (int) nRecordLength );
     }
 
     nFeatures = nFileSize / nRecordLength;
